*** empty log message ***
git-svn-id: http://code.jaspersoft.com/svn/repos/jasperreports/trunk/jasperreports@571 76e25fc1-081a-0410-bb1b-c96bc70638b7

+++ src/net/sf/jasperreports/engine/export/TextRenderer.java
@@ -167,12 +167,12 @@ public class TextRenderer
 			}
 			case JRAlignment.VERTICAL_ALIGN_MIDDLE :
 			{
-				verticalOffset = ((float)height - textHeight) / 2f;
+				verticalOffset = (height - textHeight) / 2f;
 				break;
 			}
 			case JRAlignment.VERTICAL_ALIGN_BOTTOM :
 			{
-				verticalOffset = (float)height - textHeight;
+				verticalOffset = height - textHeight;
 				break;
 			}
 			default :
